revert: cacheable board read service added.
