--- src/main/java/com/mide/gangsaeng/board/BoardReadService.java
@@ -1,5 +1,0 @@
-package com.mide.gangsaeng.board;
-
-public interface BoardReadService {
-    BoardResponse read(long id);
-}
